Fix regression with late-evaluation relationships no longer working for filters
Regression was introduced in 47080ae (PR #1348) where `is_hybrid_property` was changed to support remote hybrid properties. When lambdas were encountered in the iteration they were passed as-is to the function `get_hybrid_properties`, which in turn threw the exception `sqlalchemy.exc.NoInspectionAvailable: No inspection system is available for object of type <class 'function'>`.

Test has been updated to ensure that flask-admin will continue to support lambdas as a relationship’s first argument.